reorganize code and add more docs
