+++ crates/bevy_text/src/lib.rs
@@ -1,3 +1,33 @@
+//! This crate provides the tools for positioning and rendering text in Bevy.
+//!
+//! # `Font`
+//!
+//! Fonts contain information for drawing glyphs, which are shapes that typically represent a single character,
+//! but in some cases part of a "character" (grapheme clusters) or more than one character (ligatures).
+//!
+//! A font *face* is part of a font family,
+//! and is distinguished by its style (e.g. italic), its weight (e.g. bold) and its stretch (e.g. condensed).
+//!
+//! In Bevy, [`Font`]s are loaded by the [`FontLoader`](FontLoader) as assets,
+//! or they can be loaded as system fonts through [`TextPipeline::load_system_fonts`].
+//!
+//! # `TextPipeline`
+//!
+//! The [`TextPipeline`] resource does all of the heavy lifting for rendering text.
+//!
+//! [`Text`](Text) is first measured by creating a [`TextMeasureInfo`] in [`TextPipeline::create_text_measure`],
+//! which is called by a system.
+//!
+//! Note that text measurement is only relevant in a UI context.
+//!
+//! With the actual text bounds defined, another system passes it into [`TextPipeline::queue_text`], which:
+//!
+//! 1. creates a [`Buffer`](cosmic_text::Buffer) from the [`TextSection`]s, generating new [`FontAtlasSet`]s if necessary.
+//! 2. iterates over each glyph in the [`Buffer`](cosmic_text::Buffer) to create a [`PositionedGlyph`],
+//!    retrieving glyphs from the cache, or rasterizing to a [`FontAtlas`](FontAtlas) if necessary.
+//! 3. [`PositionedGlyph`]s are stored in a [`TextLayoutInfo`],
+//! which contains all the information that downstream systems need for rendering.
+
 #![allow(clippy::type_complexity)]
 
 mod error;
@@ -5,6 +35,7 @@ mod font;
 mod font_atlas;
 mod font_atlas_set;
 mod font_loader;
+mod glyph;
 mod pipeline;
 mod text;
 mod text2d;
@@ -14,6 +45,7 @@ pub use font::*;
 pub use font_atlas::*;
 pub use font_atlas_set::*;
 pub use font_loader::*;
+pub use glyph::*;
 pub use pipeline::*;
 pub use text::*;
 pub use text2d::*;
